Shaders: test MeshVisualizer TBN + vertex and object ID visualization.
Fails spectacularly -- the geometry shader is aware only of instanced
object ID, and of vertex ID not at all. The former was a corner case
TODO while I was adding non-instanced object ID visualization in
033e56ec23bd1a160a5af96dd8856ef92969daf1 and which I kinda forgot about,
the latter was discovered while trying to fix the former.

Fix in the next commit.