GL: improve ANGLE detection on WebGL.
It's quite easy if WEBGL_debug_renderer_info is present, but if it isn't
then our only hope is to check for ANGLE-specific extensions. Which may
cause some false positives, but since we need to enable various
workarounds for ANGLE to avoid issues, it's better than having the
detection too conservative.

The original ANGLE detection based on line size is probably not so
relevant anymore, but let's keep it there.