[wip] Trade: rebuild AbstractImageConverter around ImageData.
And not views, as ImageData contains extra metadata. Currently just the
importer state, but eventually that will be changed to a generic
key/value store for stuff like color correction, EXIF, gamma, etc., and
if the plugin interfaces would consume only views (which don't and won't
have such key/value data), the data would get lost.

A similar approach is being taken with the AbstractSceneConverter.

TODO: implement this to be more than just a TODO