[wip] Trade: un-combinatorially-explode ImageConverterFeatures.
Unlike most other options, where a plugin can support 3D but not 2D or
compressed but not uncompressed data, the "levels" option is orthogonal
to the rest -- if a format supports multi-level data, it obviously also
supports a case where there is just a single level, and if it supports
multi-level images, then it likely supports them for 1D, 2D and 3D, not
just some.

The bit pattern actually followed this reasoning already, but for some
reason I still went ahead and stamped out all possible combinations.

TODO: finish
TODO: OTOH, the ToData is probably better to not have a single bit
  common for all, as it makes it impossible to say that e.g. 3D is just
  ToFile and 2D and 1D is ToData as well (e.g. when the 3D slices would
  need multiple files). So change that to different bits instead.