Math: return an array reference instead of a pointer from data().
This makes it possible to conveniently do things like

    Containers::StridedArrayView1D<Float> array = …;
    Vector4 vector{NoInit};
    Utility::copy(array, vector); // or the other way around

which is especially useful together with the new JSON classes. In some
cases this means the function is no longer constexpr, but those weren't
constexpr because it was useful for anything, they were only because it
was possible. So this breakage shouldn't do any harm I think.