AnyImage{Importer,Converter}: detect also KTX1.
Even though KtxImporter / KtxImageConverter doesn't support these (and
probably never will), the rationale here is to provide a somewhat better
message than "unable to detect file format" when trying to open a *.ktx
file, or when trying to save to a *.ktx by accident, instead of *.ktx2
(I do that quite often). The concrete plugins are able to provide a much
better error message about version 1 not supported.