ObjImporter: support quads.
Now I'm able to measure -- with the Rungholt scene from McGuire
archives, it takes it a whooping 7.3 seconds, measured with

    magnum-sceneconverter --info-meshes --profile rungholt.obj

Compared to that, AssimpImporter in the default configuration takes 10.3
seconds, but with

    magnum-sceneconverter --info-meshes --profile -I AssimpImporter \
      -i postprocess/JoinIdenticalVertices=false,postprocess/SortByPType=false \
      rungholt.obj

it takes just 4.3 seconds. And has the same vertex data size in total,
so that should probably be the default. I got some work to do, then.