GL: deprecate list-taking Shader::compile() and ShaderProgram::link().
There's no reason for those to exist anymore -- origiinally they were
added in a hopeful attempt to make use of parallel shader compilation,
but in practice that meant compiling at most two or three shaders at
once and still stalling until that was done, so not that great at all.
The new APIs provide much better opportunities for parallelism.

Fun fact:

    CORRADE_INTERNAL_ASSERT_OUTPUT(vert.compile() && frag.compile());

is actually one character shorter than

    CORRADE_INTERNAL_ASSERT_OUTPUT(GL::Shader::compile({vert, frag}));

so not even typing convenience would be a reason to keep these.