*converter: make --plugin-dir accept the actual root plugin directory.
For some reason, it was adding also the magnum / magnum-d directory,
which isn't really useful, especially in cases where the directory is
not at all or it's desired to pick a debug plugin from a release
executable and vice versa.

Also the distancefieldconverter was still attempting to join with an
absolute path -- somehow 7fb63a9434b72c1e9da74eeb331676ebbcae0430 missed
this one.

I really need to write regression tests for all this, sigh.