Don't take an Iterable by value.
It's four pointers, twice as much as what would be acceptable. Not sure
why this happened, maybe because all those cases used an ArrayView
before and so I just changed the type without considering the difference
in its size?

Unfortunately this change also means a bump in the plugin interface
string, thus all scene converter plugins have to be updated as well.