Shaders: take constructor arguments via a Configuration instance.
Because it was no longer bearable with three UnsignedInt arguments in a
row, especially when some of them are only available on a subset of
platforms. And it would get even worse with introduction of planned
features such as multiview or skinning.

Backwards compatibility is in place, as always. To ensure nothing
breaks, this commit still has all tests and snippets using the old API.