Trade: implement support for Weights & JointIds in MeshData.
This is the first builtin array attribute, with one of the objectives
being an ability to support an arbitrary count of per-vertex weights in
a single contiguous attribute without the complexity of having to go
through several four-component attributes.

On the shader side it still needs to get cut into at most four
components per attribute, but there's no reason for such limitation to
get propagated here as well.

Co-authored-by: Vladimír Vondruš <mosra@centrum.cz>