Shaders: doc++
A considerable chunk of the docs mentioned that there has to be one
ProjectionUniform3D per draw. Probably a copypaste error from the case
where there's a combined TrannsformationProjectionUniform3D, which *is*
one per draw. Sorry for the confusion.

There was also quite a lot of documentation content referencing the old
deprecated constructors. Fixed now.