MeshTools: print a warning if compile() ignores conflicting attributes.
Which is currently the case for any two attributes of the same name
(such as two texture coordinate sets), or bitangents and object ID used
together.

Plus extending the check for matrix attributes (such as instanced
transformation conflicting with secondary joint IDs and weights).
None of these attributes are builtin though, so the check can't be
verified at the moment.

And also documenting the behavior, for some reason the comment was
apparently not updated since the MeshData2D/MeshData3D days!