DebugTools: avoid accidental OOB access in FrameProfiler internals.
The memory wasn't accessed in that case so everything is fine, but this
started tripping up the new assertions in ArrayView element access.
Fixed by moving (and thus duplicating) the access to where it is
actually needed.

Also removed the outdated comment -- there's no _currentData anymore
anywhere, not sure what was that meant to be.