Trade: improve AnimationData OOB assertion messages.
This still contained the original message which didn't say the index.
Times changed significantly since then, and the printed index really
proved to be useful.

Also fixing a OOB assertion in the test triggered by the new ArrayView
element access assertions -- the graceful assertion returns the first
element, so we have to have at least one there to return.