sceneconverter: implement a --phong-to-pbr option.
Exposes MaterialTools::phongToPbrMetallicRoughness() that got added some
time ago. Most of the code and tests is scaffolding needed for direct
material import and processing outside of the addImporterContents()
automagic, similar to what was already done for meshes and images.

In particular, adding more material conversion options such as
canonicalization or deduplication will be significantly easier now that
the basics are done and tested.