Vk: try a significantly larger count in DescriptorPool allocation test.
It was creating the pool with 7 buffers, which the driver could as well
just round up to 8 and then allocating 8 would not fail. The test used
to work on Mesa before but not anymore. So I'm creating the pool with 8
and askin for 80 which should definitely fail.

There's a lot more failures on NVidia, I "fixed" those by deleting the
NVidia Vulkan ICD file for now.