Trade: the SceneFieldData string constructors can't be constexpr.
Apparently I forgot to actually test these -- in order to fit the string
data pointer without making SceneFieldData too large, it'đ stored as an
offset from fieldData. And that's something not expressible in a
constexpr context. Thus the only way to create a constexpr string field
is by using the offset-only constructor (which is now appropriately
tested).

This change at least allowed me to move the constructor to the cpp file,
saving on header size and using more lightweight assertions.