Trade: expect MaterialData attrs to not be larger than end layer offset.
Such an unnecessary footgun -- I was already checking the other case,
having attribute data too short, but this I thought is fine because it's
not leading to any crash. Well it's leading to needless pain and
suffering, that's what it is doing!

And of course already found FIVE such bugs in just Magnum tests alone.